feat: split between components
