+++ wondrous-app/components/Icons/podIconWithoutBg.tsx
@@ -1,0 +1,15 @@
+const PodIconWithoutBg = (props) => {
+  return (
+    <svg width={35} height={35} fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="m10.554 21.712.002-.002a.01.01 0 0 1-.002.002Zm.126-.088c.049-.027.113-.057.193-.09.31-.123.782-.243 1.394-.346 1.217-.206 2.913-.335 4.797-.335s3.58.13 4.797.335c.612.103 1.084.223 1.394.347.08.032.144.062.194.09-.05.026-.113.056-.194.089-.31.124-.782.243-1.394.347-1.217.205-2.913.335-4.797.335s-3.58-.13-4.797-.335c-.612-.104-1.084-.223-1.394-.347a1.785 1.785 0 0 1-.193-.09Zm12.894.088-.002-.002a.012.012 0 0 1 .002.002Zm-.002-.174a.015.015 0 0 1 .002-.002l-.002.002Zm-13.016 0Zm1.58-1.12c1.272-.216 3.014-.347 4.928-.347 1.914 0 3.656.131 4.927.346.633.107 1.168.238 1.554.392.191.076.368.167.505.28.132.109.281.287.281.535 0 .249-.15.427-.281.536a1.913 1.913 0 0 1-.505.28c-.386.154-.92.284-1.554.391-1.271.215-3.013.346-4.927.346-1.914 0-3.656-.13-4.927-.346-.633-.107-1.168-.237-1.554-.391a1.915 1.915 0 0 1-.505-.28c-.131-.11-.281-.287-.281-.536 0-.248.15-.426.281-.535.137-.113.314-.204.505-.28.386-.154.92-.285 1.554-.392ZM10.554 17.264l.002-.002a.013.013 0 0 1-.002.002Zm.126-.088c.049-.027.113-.057.193-.09.31-.123.782-.243 1.394-.346 1.217-.206 2.913-.335 4.797-.335s3.58.129 4.797.335c.612.103 1.084.223 1.394.347.08.032.144.062.194.089-.05.027-.113.057-.194.09-.31.123-.782.243-1.394.346-1.217.206-2.913.335-4.797.335s-3.58-.129-4.797-.335c-.612-.103-1.084-.223-1.394-.346a1.802 1.802 0 0 1-.193-.09Zm12.894.088-.002-.002a.012.012 0 0 1 .002.002Zm-.002-.174.002-.002-.002.002Zm-13.016 0-.002-.002.002.002Zm1.58-1.121c1.272-.215 3.014-.346 4.928-.346 1.914 0 3.656.131 4.927.346.633.107 1.168.238 1.554.392.191.076.368.167.505.28.132.109.281.286.281.535 0 .249-.15.427-.281.536a1.916 1.916 0 0 1-.505.28c-.386.153-.92.284-1.554.391-1.271.215-3.013.346-4.927.346-1.914 0-3.656-.13-4.927-.346-.633-.107-1.168-.238-1.554-.392a1.918 1.918 0 0 1-.505-.28c-.131-.108-.281-.286-.281-.535 0-.249.15-.426.281-.535.137-.113.314-.204.505-.28.386-.154.92-.285 1.554-.392ZM10.554 12.818l.002-.002a.013.013 0 0 1-.002.002Zm.126-.088c.049-.027.113-.057.193-.09.31-.123.782-.243 1.394-.346 1.217-.206 2.913-.335 4.797-.335s3.58.13 4.797.335c.612.103 1.084.223 1.394.347.08.032.144.062.194.09-.05.026-.113.056-.194.089-.31.124-.782.243-1.394.347-1.217.205-2.913.335-4.797.335s-3.58-.13-4.797-.335c-.612-.104-1.084-.223-1.394-.347a1.802 1.802 0 0 1-.193-.09Zm12.894.088-.002-.002a.012.012 0 0 1 .002.002Zm-.002-.174.002-.002-.002.002Zm-13.016 0-.002-.002.002.002Zm1.58-1.12c1.272-.216 3.014-.347 4.928-.347 1.914 0 3.656.131 4.927.346.633.107 1.168.238 1.554.392.191.076.368.167.505.28.132.109.281.287.281.535 0 .249-.15.427-.281.536a1.916 1.916 0 0 1-.505.28c-.386.154-.92.285-1.554.391-1.271.215-3.013.346-4.927.346-1.914 0-3.656-.13-4.927-.346-.633-.106-1.168-.237-1.554-.391a1.918 1.918 0 0 1-.505-.28c-.131-.11-.281-.287-.281-.536 0-.248.15-.426.281-.535.137-.113.314-.204.505-.28.386-.154.92-.285 1.554-.392Z"
+        fill="#6ACCBC"
+      />
+      <rect x={0.566} y={0.677} width={33} height={33} rx={16.5} stroke="#6ACCBC" />
+    </svg>
+  );
+};
+
+export default PodIconWithoutBg;
